pyproject.toml
[tool.poetry]
name = "arcaflow-plugin-instructlab"
version = "0.1.0"
description = ""
authors = ["Arcalot"]
license = "Apache-2.0+GPL-2.0-only"

packages = [
   { include="instructlab_plugin.py", from="./arcaflow_plugin_instructlab"  },
]

[tool.poetry.dependencies]
python = "^3.10"
arcaflow-plugin-sdk = "^0.14.0"
instructlab-training = "^0.3.1"

[tool.poetry.dev-dependencies]
docformatter = "^1.5.0"
autoflake = "^2.0.0"
pydocstyle = "^6.1.1"
black = "^24.0.0"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
